IMA: turn ima_policy_flags into __wr_after_init
The policy flags could be targeted by an attacker to disable IMA, so
that there would be no trace of a file system modification.

Since the flags can be altered at runtime, it is not possible to make
them become fully read-only, for example with __ro_after_init

This patch makes it harder for an attacker to alter the flags.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>

CC: Andy Lutomirski <luto@amacapital.net>
CC: Nadav Amit <nadav.amit@gmail.com>
CC: Matthew Wilcox <willy@infradead.org>
CC: Peter Zijlstra <peterz@infradead.org>
CC: Kees Cook <keescook@chromium.org>
CC: Dave Hansen <dave.hansen@linux.intel.com>
CC: linux-integrity@vger.kernel.org
CC: kernel-hardening@lists.openwall.com
CC: linux-mm@kvack.org
CC: linux-kernel@vger.kernel.org