pmalloc: Partition VMALLOC space
As preparation for the protected memory allocator pmalloc(), the address
space assigned to vmalloc() is partitioned.

pmalloc will share most of vmalloc infrastructure, including the
creation of vmap areas for providing virtually contiguous memory to its
users. pmalloc will support four types of memory allocation:

* cannot be altered, once it's write protected, nor it can be released
* cannot be altered, once it's write protected, but it can be released
* can be altered, once it's write protected, and it can be released
* can be altered, once it's write protected, but it cannot be released

The reason for the differentiation is that it allows adding extra checks
to the fucntions which will perform, respectively, the release
(destruction) of areas of protected memory, and the alteration of write
protected memory.

Such functions represent a potential weakness in the protection scheme,
because they could be exploited, during an attack, for defeating the
very defences they are supposed to enforce.

In such case, even assuming the kernel had been compiled with a
CFI-enabled compiler, an ROP-based attack would still be capable of
invoking existing functions, with malicious parameters.

Example: using a write-rare function against memory that is supposed to be
         read-only.

It would be possible to tag pages somehow, for on-the-fly vetting, but
if it's assumed that the attack is capable of altering non-write
protected kernel data, then the tagging metadata would become the next
target.

A much more robust approach is to use a more intrinsic property, the
address, to classify a memory page. As bonus, it is a fairly inexpensive
test, since it only requires a handful of comparisons against the
boundaries of the expected range.

This approach is still vulnerable to attacks based on remapping, however
that vulnerability was already present and now altering the page tables
becomes a mandatory step for the attack, reducing the attack surface.

Generic read/write memory allocations provided by vmalloc() are placed
after those reserved for pmalloc().

The patch introduces the following ranges, for the purposes described:

	---------------------
	VMALLOC_RO_NO_DESTROY
	---------------------
	VMALLOC_RO_OK_DESTROY
	---------------------
	VMALLOC_WR_OK_DESTROY
	---------------------
	VMALLOC_WR_NO_DESTROY
	---------------------
	    VMALLOC_RW
	---------------------

It also adjusts accordingly various places which were previously
referring to the vmalloc address space in its entirety and instead now
should refer to the VMALLOC_RW subrange.

One downside of this approach is that now it is necessary to estimate
the size of the address space required by each type of allocation.
Doing it for a 32 bit address requires higher accuracy than for a 64 bit
one. However the allocations are likely to be mostly permanent and
genarate little fragmentation in the address ranges.

Special consideration must be paid to never-shrinking scenarios, like
IMA, for example, where the allocations are never released and will
eventually use up all the address space. This is not a new problem, as
IMA has the potential to use up all the address space even without this
patch, but it is now necessary to estimate the number of entries that
will be generated during the expected uptime.

XXX Note: it might be necessary to support configuring the size of each
range through the command line. Furthermore, it might be desirable to
configure either at build time or through the command line, which range
to use for a certain purpose. Example: SELinux on Android will not
reload its policy DB, therefore it should be kept in
VMALLOC_RO_NO_DESTROY, instead of VMALLOC_RO_OK_DESTROY, as a generic
distro might have to do.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>